Update README
There was a young flower in the desert where all was dry and sad looking…… It was growing by itself…… enjoying every day…… and saying to the sun "When shall I be grown up？" And the sun would say "Be patient - Each time I touch you， you grow a little……" She was so pleased. Because she would have a chance to bring beauty to this corner of sand…… And this is all she wanted to do - bring a little bit of beauty to this world.